requirements.txt
langchain==0.2.12
langchain_community==0.2.11
langchain_core==0.2.29
langchain_google_genai==1.0.8
python-dotenv==1.0.1
streamlit==1.37.1
streamlit_chat==0.1.1
